+++ src/test/java/LadderPrintTest.java
@@ -1,7 +1,7 @@
-import LadderMaker.LadderMaker;
+import ladderMaker.LadderMaker;
 import domain.LadderNumber;
 import game.LadderGame;
-import LadderMaker.Liner;
+import ladderMaker.Liner;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
